build: init templ
