+++ Dockerfile
@@ -31,7 +31,7 @@ ARG TARGETARCH
 # source code into the container.
 RUN --mount=type=cache,target=/go/pkg/mod/ \
     --mount=type=bind,target=. \
-    CGO_ENABLED=0 GOARCH=$TARGETARCH go build -o /bin/server cmd/api/main.go
+    CGO_ENABLED=0 GOARCH=$TARGETARCH go build -o /bin/server cmd/main.go
 
 FROM alpine:latest AS final
