+++ Makefile
@@ -1,36 +1,13 @@
-# Simple Makefile for a Go project
-# Build the application
-all: build
-
 build:
-	@echo "Building..."
-	@go build -o main cmd/api/main.go
+	@echo "Building application..."
+	@go build -o main cmd/main.go
 
-# Run the application
 run:
-	@go run cmd/api/main.go
+	@go run cmd/main.go
 
-# Clean the binary
 clean:
 	@echo "Cleaning..."
 	@rm -f main
-
-# Live Reload
-watch:
-	@if command -v air > /dev/null; then \
-		(air & templ generate --watch & ./tailwindcss -i ./static/css/input.css -o ./static/css/style.css --watch & wait); \
-		echo "Watching...";\
-	else \
-		read -p "Go's 'air' is not installed on your machine. Do you want to install it? [Y/n] " choice; \
-		if [ "$$choice" != "n" ] && [ "$$choice" != "N" ]; then \
-			go install github.com/air-verse/air@latest; \
-			(air & templ generate --watch & wait); \
-			echo "Watching...";\
-		else \
-			echo "You chose not to install air. Exiting..."; \
-			exit 1; \
-		fi; \
-	fi
 
 migrate-up:
 	@echo "Migrating up..."
@@ -77,3 +54,18 @@ docker-db-restart:
 docker-delete:
 	@echo "Deleting Docker containers..."
 	@docker compose down --volumes --networks --remove-orphans
+
+build-templ:
+	@templ generate
+
+build-tailwind:
+	./tailwindcss -m -i ./public/tailwind.css -o ./public/styles.css $(ARGS)
+
+dev-server:
+	@air
+
+dev-templ:
+	@templ generate --watch
+
+dev-tailwind:
+	@make ARGS="--watch" build-tailwind
