+++ internal/utils/env.go
@@ -52,7 +52,7 @@ func GetServerAddress() string {
 func LoadEnv() {
 	err := godotenv.Load(".env")
 	if err != nil {
-		log.Fatal("Error loading .env file")
+		log.Fatal("Error loading .env file", err)
 	}
 
 	for key, value := range env {
